+++ tests/fixtures/error_response.json
@@ -1,0 +1,6 @@
+{
+  "message": "",
+  "error": "",
+  "error_description": "",
+  "errorcode": 0
+}
