+++ tests/fixtures/get_list_response.json
@@ -1,0 +1,30 @@
+{
+  "uuid": "00000000-0000-0000-0000-000000000000",
+  "status": "SHARED",
+  "items": {
+    "purchase": [
+      {
+        "uuid": "43bdd5a2-740a-4230-8b27-d0bbde886da7",
+        "specification": "grün",
+        "itemId": "Paprika"
+      },
+      {
+        "uuid": "2de9d1c0-c211-4129-b6c5-c1260c3fc735",
+        "specification": "gelb",
+        "itemId": "Zucchetti"
+      }
+    ],
+    "recently": [
+      {
+        "uuid": "5681ed79-c8e4-4c8b-95ec-112999d016c0",
+        "specification": "rot",
+        "itemId": "Paprika"
+      },
+      {
+        "uuid": "01eea2cd-f433-4263-ad08-3d71317c4298",
+        "specification": "",
+        "itemId": "Pouletbrüstli"
+      }
+    ]
+  }
+}
