+++ tests/fixtures/load_lists_response.json
@@ -1,0 +1,9 @@
+{
+  "lists": [
+    {
+      "listUuid": "00000000-0000-0000-0000-000000000000",
+      "name": "Einkauf",
+      "theme": "ch.publisheria.bring.theme.home"
+    }
+  ]
+}
