+++ tests/fixtures/login_response.json
@@ -1,0 +1,12 @@
+{
+  "uuid": "00000000-0000-0000-0000-000000000000",
+  "publicUuid": "00000000-0000-0000-0000-000000000000",
+  "email": "EMAIL",
+  "name": "NAME",
+  "photoPath": "",
+  "bringListUUID": "00000000-0000-0000-0000-000000000000",
+  "access_token": "ACCESS_TOKEN",
+  "refresh_token": "REFRESH_TOKEN",
+  "token_type": "Bearer",
+  "expires_in": 604799
+}
