+++ tests/fixtures/token_response.json
@@ -1,0 +1,6 @@
+{
+  "access_token": "{access_token}",
+  "refresh_token": "{refresh_token}",
+  "token_type": "Bearer",
+  "expires_in": 604799
+}
