+++ tests/fixtures/user_account_response.json
@@ -1,0 +1,17 @@
+{
+  "userUuid": "00000000-0000-0000-0000-000000000000",
+  "publicUserUuid": "00000000-0000-0000-0000-000000000000",
+  "email": "{email}",
+  "emailVerified": true,
+  "name": "{user_name}",
+  "photoPath": "bring/user/portrait/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxx",
+  "userLocale": { "language": "de", "country": "DE" },
+  "premiumConfiguration": {
+    "hasPremium": false,
+    "hideSponsoredProducts": false,
+    "hideSponsoredTemplates": false,
+    "hideSponsoredPosts": false,
+    "hideSponsoredCategories": false,
+    "hideOffersOnMain": false
+  }
+}
