+++ tests/test_load_lists.py
@@ -1,0 +1,98 @@
+"""Tests for load_lists method."""
+
+import asyncio
+from http import HTTPStatus
+
+import aiohttp
+from aioresponses import aioresponses
+import pytest
+from syrupy.assertion import SnapshotAssertion
+
+from bring_api import (
+    Bring,
+    BringAuthException,
+    BringParseException,
+    BringRequestException,
+)
+
+from .conftest import UUID, load_fixture
+
+
+@pytest.mark.usefixtures("mocked")
+async def test_load_lists(
+    bring: Bring,
+    snapshot: SnapshotAssertion,
+) -> None:
+    """Test load_lists."""
+    await bring.login()
+    lists = await bring.load_lists()
+
+    assert lists == snapshot
+
+
+@pytest.mark.parametrize(
+    ("status", "exception"),
+    [
+        (HTTPStatus.OK, BringParseException),
+        (HTTPStatus.UNAUTHORIZED, BringAuthException),
+    ],
+)
+async def test_parse_exception(
+    mocked: aioresponses,
+    bring: Bring,
+    status: HTTPStatus,
+    exception: type[Exception],
+) -> None:
+    """Test parse exceptions."""
+    await bring.login()
+    mocked.clear()
+    mocked.get(
+        f"https://api.getbring.com/rest/bringusers/{UUID}/lists",
+        status=status,
+        body="not json",
+        content_type="application/json",
+    )
+
+    with pytest.raises(exception):
+        await bring.load_lists()
+
+
+async def test_unauthorized(
+    mocked: aioresponses,
+    bring: Bring,
+) -> None:
+    """Test unauthorized exception."""
+    await bring.login()
+    mocked.clear()
+    mocked.get(
+        f"https://api.getbring.com/rest/bringusers/{UUID}/lists",
+        status=HTTPStatus.UNAUTHORIZED,
+        body=load_fixture("error_response.json"),
+    )
+
+    with pytest.raises(BringAuthException):
+        await bring.load_lists()
+
+
+@pytest.mark.parametrize(
+    "exception",
+    [
+        asyncio.TimeoutError,
+        aiohttp.ClientError,
+    ],
+)
+async def test_request_exception(
+    mocked: aioresponses,
+    bring: Bring,
+    exception: type[Exception],
+) -> None:
+    """Test request exceptions."""
+    await bring.login()
+    mocked.clear()
+    mocked.get(
+        f"https://api.getbring.com/rest/bringusers/{UUID}/lists",
+        exception=exception,
+    )
+
+    with pytest.raises(BringRequestException):
+        await bring.load_lists()
